Move client to connect to server to proper destination
